# docs/10-client-api/Client-API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EDUagent Client-API",
    "version": "5.0.0",
    "description": "前端 (Vue3) ↔ Backend (FastAPI, Port 8001) 接口规范。\n统一返回格式: {code, message, data}。\n流式接口使用 SSE (text/event-stream)。\n前端发起刷新/生成接口后收到 task_id，并通过 /tasks/{task_id} 轮询；Backend 负责创建任务、记录归属用户和任务类型；Agent Service 长任务完成后由 Backend 接收 webhook 回调，前端不直接等待 webhook。"
  },
  "servers": [
    {
      "url": "http://localhost:8001/api/v1",
      "description": "本地开发环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "注册",
        "description": "使用永久注册码+邮箱+密码+用户名+算术验证码注册。注册码区分角色：student 注册学生账号，teacher 注册教师账号；admin 不开放注册，由系统预置。student / teacher 为内置永久码，可重复使用。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "created"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RegisterData"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "注册码无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "邮箱或用户名冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "登录",
        "description": "邮箱+密码+算术验证码登录，返回单个 JWT token。v1 开发阶段暂不实现 refresh_token 轮换。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginData"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "邮箱或密码错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "type": "null"
          }
        }
      },
...
      "LoginData": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT 访问令牌"
          },
          "expires_in": {
            "type": "integer",
            "description": "token 有效期（秒），开发阶段可设置为 7 天"
          },
          "user": {
            "$ref": "#/components/schemas/LoginUser"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "captcha_token",
          "captcha_code"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "邮箱"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "captcha_token": {
            "type": "string",
            "description": "验证码标识"
          },
          "captcha_code": {
            "type": "string",
            "description": "验证码答案"
          }
        }
      },
      "LoginUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "description": "student / teacher / admin"
          }
        }
      },
...
      "RegisterData": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "用户 ID"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "registration_code",
          "email",
          "password",
          "username",
          "captcha_token",
          "captcha_code"
        ],
        "properties": {
          "registration_code": {
            "type": "string",
            "description": "永久注册码：student 注册学生账号，teacher 注册教师账号；student / teacher 为内置永久码，可重复使用"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "邮箱地址"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 32,
            "description": "密码（8-32 位，含大小写字母+数字）"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "description": "用户名"
          },
          "real_name": {
            "type": "string",
            "description": "真实姓名，注册成功后写入用户基础资料；不得传入 Agent Service"
          },
          "student_id": {
            "type": "string",
            "description": "学号或工号，注册成功后写入用户基础资料；不得传入 Agent Service"
          },
          "major": {
            "type": "string",
            "description": "专业，可作为脱敏学习上下文"
          },
          "grade": {
            "type": "string",
            "description": "年级，可作为脱敏学习上下文"
          },
          "guidance_level": {
            "type": "string",
            "enum": [
              "L1",
              "L2",
              "L3"
            ],
            "default": "L2",
            "description": "全局引导粒度，可作为脱敏学习上下文"
          },
          "captcha_token": {
            "type": "string",
            "description": "算术验证码标识"
          },
          "captcha_code": {
            "type": "string",
            "description": "算术验证码答案"
          }
        }
      },
...
    }
  }
}